+++ CHANGELOG.md
@@ -7,6 +7,11 @@
 * Remove bigtests from github actions. Refs UICIRC-1079.
 * Create a new setting - Enable ECS title level requests. Refs UICIRC-1040.
 * Add view permissions for Circulation settings. Refs UICIRC-1056.
+* Create new permission 'Settings (Circ): Can enable request print details'. Refs UICIRC-1093.
+* Create a new circ settings page and configuration to enable request print history. Refs UICIRC-1087.
+* Add page title to Circulation settings - View Print Details Page. Refs UICIRC-1089.
+* Add staffUsername as staff slips token in Settings. Refs UICIRC-1101.
+* Add request.barcodeImage as staff slips token in Settings. Refs UICIRC-1103.
 
 ## [9.1.0](https://github.com/folio-org/ui-circulation/tree/v9.1.0) (2024-03-22)
 [Full Changelog](https://github.com/folio-org/ui-circulation/compare/v9.0.5...v9.1.0)